refactor(SqlSchemaUpdater): Extract namespace, table, and column checks to indiv classes.
